+++ packages/sdk/CHANGELOG.md
@@ -22,7 +22,7 @@
 
 - Updated dependencies [973858c6]
   - @inlang/message-lint-rule@1.1.0
-  - @inlang/project-config@1.2.0
+  - @inlang/project-settings@1.2.0
   - @inlang/language-tag@1.1.0
   - @inlang/translatable@1.1.0
   - @inlang/message@1.1.0